+++ database/migrations/2023_03_27_201126_create_payments_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
     {
         Schema::create('payments', function (Blueprint $table) {
             $table->id();
-            $table->integer('order_id');
+            $table->unsignedBigInteger('order_id');
             $table->foreign('order_id')->references('id')->on('orders');
             $table->enum('method', ['cash', 'card']);
             $table->timestamps();
